Finished parser implementation, further refactored lexer, removed 'dominates' keyword and terminals
The 'rulebase' parser has been completed and is ready for testing. To allow for a smoother and easier testing process I refactored the lexer to its own function (createLexer) and also extended it to allow for comments: marked with # and anything following it on the same line will be treated as such, note the line need not start with the # symbol. Finally, the 'dominates' keyword along with the terminals have been removed, however regarding the latter certain elements in the code are retained should the need for such functionality arise in the future.